fix(render): address perf docs and frame timing review
