+++ packages/render/src/pipeline.ts
@@ -1,53 +1,37 @@
-import { Shader, VGPUError, type Device } from "@vgpu/core";
+import { VGPUError, type Device } from "@vgpu/core";
+import { toRenderPipelineDescriptor } from "./pipeline-descriptor.ts";
+import type { RenderPipelineAsyncFallback, RenderPipelineOptions } from "./pipeline-descriptor.ts";
 
-export type RenderPipelineShaderInput = Shader | GPUShaderModule;
-export type RenderPipelineAsyncFallback = "sync" | "throw";
-
-export interface RenderPipelineStageOptions {
-  /** VGPU Shader or raw GPUShaderModule for this stage. Defaults to RenderPipelineOptions.shader. */
-  readonly shader?: RenderPipelineShaderInput;
-  /** Descriptor-like alias for shader. */
-  readonly module?: RenderPipelineShaderInput;
-  /** Entry-point name. Kept for backwards compatibility with the first render helper API. */
-  readonly entry?: string;
-  /** Descriptor-like entry-point name. */
-  readonly entryPoint?: string;
-  readonly constants?: Record<string, GPUPipelineConstantValue>;
-}
-
-export interface RenderPipelineVertexOptions extends RenderPipelineStageOptions {
-  readonly buffers?: readonly (GPUVertexBufferLayout | null)[];
-}
-
-export interface RenderPipelineFragmentOptions extends RenderPipelineStageOptions {
-  readonly targets: readonly (GPUColorTargetState | null)[];
-}
-
-export interface RenderPipelineOptions {
-  /** Optional shared shader module used by vertex/fragment stages that do not provide their own module. */
-  readonly shader?: RenderPipelineShaderInput;
-  readonly vertex: RenderPipelineVertexOptions;
-  readonly fragment?: RenderPipelineFragmentOptions;
-  readonly primitive?: GPUPrimitiveState;
-  readonly depthStencil?: GPUDepthStencilState;
-  readonly multisample?: GPUMultisampleState;
-  readonly layout?: GPUPipelineLayout | "auto";
-  readonly label?: string;
-  /** createRenderPipelineAsync fallback when GPUDevice.createRenderPipelineAsync is unavailable. Defaults to "sync". */
-  readonly fallback?: RenderPipelineAsyncFallback;
-}
+export type {
+  RenderPipelineAsyncFallback,
+  RenderPipelineFragmentOptions,
+  RenderPipelineOptions,
+  RenderPipelineShaderInput,
+  RenderPipelineStageOptions,
+  RenderPipelineVertexOptions,
+} from "./pipeline-descriptor.ts";
 
 // Warn once per JS process. This keeps compatibility fallback visible without spamming
 // apps that intentionally warm up multiple pipelines on implementations without
 // GPUDevice.createRenderPipelineAsync().
 let didWarnAboutAsyncFallback = false;
 
+type AsyncPipelineWhere = "createRenderPipelineAsync" | "createRenderPipelineFromDescriptorAsync";
+
 export function createRenderPipeline(device: Device, opts: RenderPipelineOptions): GPURenderPipeline {
   return device.gpu.createRenderPipeline(toRenderPipelineDescriptor(opts));
 }
 
-export async function createRenderPipelineAsync(device: Device, opts: RenderPipelineOptions): Promise<GPURenderPipeline> {
-  return createPipelineAsync(device, toRenderPipelineDescriptor(opts), opts.fallback);
+export async function createRenderPipelineAsync(
+  device: Device,
+  opts: RenderPipelineOptions,
+): Promise<GPURenderPipeline> {
+  return createPipelineAsync(
+    device,
+    toRenderPipelineDescriptor(opts),
+    opts.fallback,
+    "createRenderPipelineAsync",
+  );
 }
 
 /**
@@ -58,7 +42,10 @@ export async function createRenderPipelineAsync(device: Device, opts: RenderPipe
  * `Device` wrapper to forward it — no `RenderPipelineOptions` reshape. The
  * descriptor is passed through unchanged.
  */
-export function createRenderPipelineFromDescriptor(device: Device, descriptor: GPURenderPipelineDescriptor): GPURenderPipeline {
+export function createRenderPipelineFromDescriptor(
+  device: Device,
+  descriptor: GPURenderPipelineDescriptor,
+): GPURenderPipeline {
   return device.gpu.createRenderPipeline(descriptor);
 }
 
@@ -76,13 +63,14 @@ export async function createRenderPipelineFromDescriptorAsync(
   descriptor: GPURenderPipelineDescriptor,
   fallback?: RenderPipelineAsyncFallback,
 ): Promise<GPURenderPipeline> {
-  return createPipelineAsync(device, descriptor, fallback);
+  return createPipelineAsync(device, descriptor, fallback, "createRenderPipelineFromDescriptorAsync");
 }
 
 async function createPipelineAsync(
   device: Device,
   descriptor: GPURenderPipelineDescriptor,
   fallback: RenderPipelineAsyncFallback | undefined,
+  where: AsyncPipelineWhere,
 ): Promise<GPURenderPipeline> {
   const createAsync = device.gpu.createRenderPipelineAsync;
   if (typeof createAsync === "function") {
@@ -94,60 +82,12 @@ async function createPipelineAsync(
       code: "VGPU-RENDER-PIPELINE-ASYNC-UNAVAILABLE",
       message: "GPUDevice.createRenderPipelineAsync is unavailable on this WebGPU implementation.",
       fix: "Use fallback: 'sync' for compatibility or call createRenderPipeline() explicitly during setup/warmup.",
-      where: "createRenderPipelineAsync",
+      where,
     });
   }
 
   warnAsyncFallbackOnce();
   return device.gpu.createRenderPipeline(descriptor);
-}
-
-function toRenderPipelineDescriptor(opts: RenderPipelineOptions): GPURenderPipelineDescriptor {
-  const descriptor: GPURenderPipelineDescriptor = {
-    label: opts.label,
-    layout: opts.layout ?? "auto",
-    vertex: {
-      module: stageModule(opts.vertex, opts.shader, "vertex"),
-      entryPoint: stageEntryPoint(opts.vertex),
-      constants: opts.vertex.constants,
-      buffers: opts.vertex.buffers ? [...opts.vertex.buffers] : undefined,
-    },
-    primitive: opts.primitive,
-    depthStencil: opts.depthStencil,
-    multisample: opts.multisample,
-  };
-
-  if (opts.fragment) {
-    descriptor.fragment = {
-      module: stageModule(opts.fragment, opts.shader, "fragment"),
-      entryPoint: stageEntryPoint(opts.fragment),
-      constants: opts.fragment.constants,
-      targets: [...opts.fragment.targets],
-    };
-  }
-
-  return descriptor;
-}
-
-function stageModule(stage: RenderPipelineStageOptions, fallback: RenderPipelineShaderInput | undefined, stageName: "vertex" | "fragment"): GPUShaderModule {
-  const module = stage.module ?? stage.shader ?? fallback;
-  if (!module) {
-    throw new VGPUError({
-      code: "VGPU-RENDER-PIPELINE-MISSING-SHADER",
-      message: `Missing shader module for ${stageName} stage.`,
-      fix: "Pass options.shader for a shared module or pass vertex/fragment module or shader.",
-      where: `createRenderPipeline.${stageName}`,
-    });
-  }
-  return isShader(module) ? module.gpu : module;
-}
-
-function stageEntryPoint(stage: RenderPipelineStageOptions): string | undefined {
-  return stage.entryPoint ?? stage.entry;
-}
-
-function isShader(input: RenderPipelineShaderInput): input is Shader {
-  return input instanceof Shader;
 }
 
 export function __resetCreateRenderPipelineAsyncFallbackWarningForTests(): void {
