+++ packages/render/tests/perf.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest";
 import { createNodeAdapter } from "@vgpu/adapter-node";
-import { App, type Device, type Texture } from "@vgpu/core";
+import { App, Device, createMockGPUDevice, type Texture } from "@vgpu/core";
 import { beginFrame, createRenderPipeline } from "@vgpu/render";
 import { gpuFrameTime, pixelDiff } from "@vgpu/render/perf";
 
@@ -34,6 +34,24 @@ describe("pixelDiff", () => {
   test("a length mismatch surfaces as a maximal difference", async () => {
     const got = await pixelDiff(new Uint8Array([1, 2]), new Uint8Array([1, 2, 3, 4]));
     expect(got.maxByte).toBe(255);
+  });
+});
+
+describe("gpuFrameTime", () => {
+  test("wall-clock fallback reports one sample per measured frame", async () => {
+    const device = new Device(createMockGPUDevice(), null);
+    const indexes: number[] = [];
+
+    const result = await gpuFrameTime(
+      device,
+      (_frame, i) => indexes.push(i),
+      { frames: 5, warmup: 2, batch: 2, forceWallClock: true },
+    );
+
+    expect(result.method).toBe("wall-clock");
+    expect(result.samples).toBe(5);
+    expect(indexes).toEqual([0, 1, 0, 1, 2, 3, 4]);
+    device.destroy();
   });
 });
